+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - NSR quay info mapped into the Calls table.
+- All timestamp columns converted to datetime.
 
 ## [0.1.0] – 2024-04-18
 ### Added